+++ src/polhemus_ros_driver/viper.cpp
@@ -42,7 +42,7 @@
 }
 
 Viper::Viper(std::string name, uint16_t rx_buffer_size, uint16_t tx_buffer_size,
-             int sensors_right_glove, int sensors_left_glove)
+             uint8_t sensors_right_glove, uint8_t sensors_left_glove)
     : Polhemus(name, rx_buffer_size, tx_buffer_size, sensors_right_glove, sensors_left_glove)
 {
 }
